backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.7
anthropic==0.7.8
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
aiofiles==23.2.1
python-dotenv==1.0.0